backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# Vector DB
chromadb==0.5.0

# Embeddings
sentence-transformers==3.0.0
torch==2.3.0
transformers==4.41.0

# OCR
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
pypdf2==3.0.1
python-docx==1.1.0

# Speech-to-text
openai-whisper

# LLM (Ollama client)
ollama==0.2.1
httpx==0.27.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
aiofiles==23.2.1
python-magic==0.4.27
tqdm==4.66.4
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0

# Reranker
# cross-encoder handled via sentence-transformers

# Export
reportlab==4.2.0
openpyxl==3.1.4

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
